feat: add brokre binary and improve injector and SSH identity handling
- Introduced a new binary file `brokre` for enhanced functionality.
- Updated `parent_pid_of` function in `injector.rs` to return the parent process ID directly.
- Refactored logic in `pty_session.rs` to streamline field injection process.
- Improved condition checks in `ssh_identity.rs` for better argument handling.

+++ src/cli/injector.rs
@@ -191,7 +191,7 @@ fn parent_pid_of(pid: u32) -> Option<u32> {
     if rc <= 0 {
         return None;
     }
-    Some(info.pbi_ppid as u32)
+    Some(info.pbi_ppid)
 }
 
 #[cfg(all(unix, target_os = "linux"))]
